feat: extract svgr as a standalone package
