+++ .changeset/blue-mugs-refuse.md
@@ -2,4 +2,4 @@
 '@rsbuild/plugin-react': patch
 ---
 
-feat(plugin-react): extract antd/arco/splitChunks/svgr logics into react plugin
+feat(plugin-react): extract antd/arco/splitChunks logics into react plugin
