+++ .changeset/silent-windows-add.md
@@ -1,0 +1,5 @@
+---
+'@rsbuild/plugin-svgr': patch
+---
+
+feat: extract svgr as a standalone package
