+++ e2e/cases/svg/svg-url/src/App.jsx
@@ -1,5 +1,5 @@
 // eslint-disable-next-line no-unused-vars
-import { ReactComponent as Logo } from './app.svg';
+import Logo from './app.svg';
 import svgImg from './app.svg?url';
 import './App.css';
