+++ e2e/cases/svg/svg.test.ts
@@ -2,10 +2,11 @@ import { join } from 'path';
 import { expect, test } from '@playwright/test';
 import { build, getHrefByEntryName } from '@scripts/shared';
 import { pluginReact } from '@rsbuild/plugin-react';
+import { pluginSvgr } from '@rsbuild/plugin-svgr';
 
 const fixtures = __dirname;
 
-test('svg', async ({ page }) => {
+test('svg (assets)', async ({ page }) => {
   const rsbuild = await build({
     cwd: join(fixtures, 'svg-assets'),
     entry: {
@@ -33,14 +34,19 @@ test('svg', async ({ page }) => {
   rsbuild.close();
 });
 
-test('svg (default)', async ({ page }) => {
+test('svgr (defaultExport url)', async ({ page }) => {
   const rsbuild = await build({
     cwd: join(fixtures, 'svg'),
     entry: {
       main: join(fixtures, 'svg', 'src/index.js'),
     },
     runServer: true,
-    plugins: [pluginReact()],
+    plugins: [
+      pluginReact(),
+      pluginSvgr({
+        svgDefaultExport: 'url',
+      }),
+    ],
   });
 
   await page.goto(getHrefByEntryName('main', rsbuild.port));
@@ -75,7 +81,7 @@ test('svgr (defaultExport component)', async ({ page }) => {
     },
     runServer: true,
     rsbuildConfig: {},
-    plugins: [pluginReact()],
+    plugins: [pluginReact(), pluginSvgr()],
   });
 
   await page.goto(getHrefByEntryName('main', rsbuild.port));
@@ -87,13 +93,13 @@ test('svgr (defaultExport component)', async ({ page }) => {
   rsbuild.close();
 });
 
-test('svgr (url)', async ({ page }) => {
+test('svgr (query url)', async ({ page }) => {
   const rsbuild = await build({
     cwd: join(fixtures, 'svg-url'),
     entry: {
       main: join(fixtures, 'svg-url', 'src/index.js'),
     },
-    plugins: [pluginReact()],
+    plugins: [pluginReact(), pluginSvgr()],
     runServer: true,
   });
 
@@ -124,7 +130,12 @@ test('svgr (external react)', async ({ page }) => {
       main: join(fixtures, 'svg-external-react', 'src/index.js'),
     },
     runServer: true,
-    plugins: [pluginReact()],
+    plugins: [
+      pluginReact(),
+      pluginSvgr({
+        svgDefaultExport: 'url',
+      }),
+    ],
     rsbuildConfig: {
       output: {
         externals: {
@@ -156,35 +167,6 @@ test('svgr (external react)', async ({ page }) => {
   await expect(
     page.evaluate(
       `getComputedStyle(document.getElementById('test-css')).backgroundImage.includes('url("data:image/svg')`,
-    ),
-  ).resolves.toBeTruthy();
-
-  rsbuild.close();
-});
-
-test('svgr (disableSvgr)', async ({ page }) => {
-  const rsbuild = await build({
-    cwd: join(fixtures, 'svg-assets'),
-    entry: {
-      main: join(fixtures, 'svg-assets', 'src/index.js'),
-    },
-    runServer: true,
-    plugins: [pluginReact()],
-  });
-
-  await page.goto(getHrefByEntryName('main', rsbuild.port));
-
-  // test svg asset
-  await expect(
-    page.evaluate(
-      `document.getElementById('test-img').src.includes('static/svg/app')`,
-    ),
-  ).resolves.toBeTruthy();
-
-  // test svg asset in css
-  await expect(
-    page.evaluate(
-      `getComputedStyle(document.getElementById('test-css')).backgroundImage.includes('static/svg/app')`,
     ),
   ).resolves.toBeTruthy();
