+++ e2e/cases/svg/svgo.test.ts
@@ -3,6 +3,7 @@ import { readFileSync } from 'fs';
 import { expect, test } from '@playwright/test';
 import { build } from '@scripts/shared';
 import { pluginReact } from '@rsbuild/plugin-react';
+import { pluginSvgr } from '@rsbuild/plugin-svgr';
 
 test('should preserve viewBox after svgo minification', async () => {
   const fixture = join(__dirname, 'svgo-minify-view-box');
@@ -11,7 +12,12 @@ test('should preserve viewBox after svgo minification', async () => {
     entry: {
       main: join(fixture, 'src/index.jsx'),
     },
-    plugins: [pluginReact()],
+    plugins: [
+      pluginReact(),
+      pluginSvgr({
+        svgDefaultExport: 'url',
+      }),
+    ],
   };
 
   const rsbuild = await build(buildOpts);
@@ -34,7 +40,12 @@ test('should add id prefix after svgo minification', async () => {
     entry: {
       main: join(fixture, 'src/index.jsx'),
     },
-    plugins: [pluginReact()],
+    plugins: [
+      pluginReact(),
+      pluginSvgr({
+        svgDefaultExport: 'url',
+      }),
+    ],
   };
 
   const rsbuild = await build(buildOpts);
