+++ packages/plugin-svgr/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @rsbuild/plugin-svgr
