+++ packages/plugin-svgr/modern.config.ts
@@ -1,0 +1,3 @@
+import baseConfig from '../../scripts/modern.base.config';
+
+export default baseConfig;
