+++ packages/plugin-svgr/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@rsbuild/tsconfig/base",
+  "compilerOptions": {
+    "outDir": "./dist",
+    "baseUrl": "./"
+  },
+  "include": ["src"],
+  "exclude": ["./tests/fixtures"]
+}
